functions in hub
